update-2019-12-13

+++ src/shared/reboot-util.c
@@ -6,16 +6,20 @@
 #include "alloc-util.h"
 #include "fileio.h"
 #include "log.h"
+#include "proc-cmdline.h"
 #include "raw-reboot.h"
 #include "reboot-util.h"
 #include "string-util.h"
 #include "umask-util.h"
 #include "virt.h"
 
-int update_reboot_parameter_and_warn(const char *parameter) {
+int update_reboot_parameter_and_warn(const char *parameter, bool keep) {
         int r;
 
         if (isempty(parameter)) {
+                if (keep)
+                        return 0;
+
                 if (unlink("/run/systemd/reboot-param") < 0) {
                         if (errno == ENOENT)
                                 return 0;
@@ -32,6 +36,18 @@ int update_reboot_parameter_and_warn(const char *parameter) {
                 if (r < 0)
                         return log_warning_errno(r, "Failed to write reboot parameter file: %m");
         }
+
+        return 0;
+}
+
+int read_reboot_parameter(char **parameter) {
+        int r;
+
+        assert(parameter);
+
+        r = read_one_line_file("/run/systemd/reboot-param", parameter);
+        if (r < 0 && r != -ENOENT)
+                return log_debug_errno(r, "Failed to read /run/systemd/reboot-param: %m");
 
         return 0;
 }
@@ -80,4 +96,15 @@ int reboot_with_parameter(RebootFlags flags) {
         (void) reboot(RB_AUTOBOOT);
 
         return log_full_errno(flags & REBOOT_LOG ? LOG_ERR : LOG_DEBUG, errno, "Failed to reboot: %m");
+}
+
+int shall_restore_state(void) {
+        bool ret;
+        int r;
+
+        r = proc_cmdline_get_bool("systemd.restore_state", &ret);
+        if (r < 0)
+                return r;
+
+        return r > 0 ? ret : true;
 }
